Update README.md
My project was created with the intention of provide the user with information about the countries around the world. I tried to implement all the concepts we learned in class, I ended up having a better understanding of how react and redux work together. Was an overall fun project I learned some new thing on css and react as well.